harness-kimi/project/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
